+++ lib/src/maj_builder.dart
@@ -11,12 +11,23 @@ import 'package:memory_and_json_directories/src/maj_node.dart';
 import 'package:provider/provider.dart';
 import 'package:memory_and_json_directories/src/maj_provider.dart';
 
+/// This is the widget used to display your tree in memory
 class MAJBuilder extends StatefulWidget {
+  /// The root of the tree you wish to display
+  /// this does not have to be a root node; however,
+  /// setting this as the root is the convention
   final MAJNode root;
+
+  /// the key you use to reference the map of nodes
+  /// don't set if you only intend to have one tree
+  /// set if you wish to have more than one tree in
+  /// memory at once
+  final String mapKey;
 
   const MAJBuilder({
     Key? key,
     required this.root,
+    this.mapKey = MAJProvider.defaultMapKey,
   }) : super(key: key);
 
   @override
@@ -26,15 +37,13 @@ class MAJBuilder extends StatefulWidget {
 }
 
 class _MAJBuilderState extends State<MAJBuilder> {
-  late String _currentPath;
   late MAJNode _currentNode;
 
   @override
   void initState() {
     super.initState();
 
-    // set the current path to the path of the root
-    _currentPath = widget.root.path;
+    // set the current node to the passed root reference
     _currentNode = widget.root;
   }
 
@@ -45,8 +54,8 @@ class _MAJBuilderState extends State<MAJBuilder> {
       providers: [
         ChangeNotifierProvider(
           create: (_) => MAJProvider(
-            currentPath: _currentPath,
-            currentNode: widget.root,
+            currentNode: _currentNode,
+            mapKey: widget.mapKey,
           ),
         ),
       ],
@@ -55,18 +64,9 @@ class _MAJBuilderState extends State<MAJBuilder> {
       // provider state management works
       child: Builder(
         builder: (context) {
-          // get current path
-          _currentPath = context.watch<MAJProvider>().currentPath;
-
-          // get current node
+          // get current node, and build it
           _currentNode = context.watch<MAJProvider>().currentNode;
-
-          // use the appropriate method to build the node
-          if (context.read<MAJProvider>().byPathElseByNode) {
-            return MAJProvider.map[_currentPath]!.build(context);
-          } else {
-            return _currentNode.build(context);
-          }
+          return _currentNode.build(context);
         },
       ),
     );