+++ lib/src/maj_provider.dart
@@ -10,18 +10,14 @@ import 'package:flutter/material.dart';
 import 'package:memory_and_json_directories/memory_and_json_directories.dart';
 
 class MAJProvider with ChangeNotifier {
-  String currentPath;
-  MAJNode currentNode;
-  bool byPathElseByNode = false;
+  static const defaultMapKey = "default";
 
   /// a map with keys == the paths of all nodes
   /// values == references to all nodes
-  static final Map<String, MAJNode> map = {};
-
-  MAJProvider({
-    required this.currentPath,
-    required this.currentNode,
-  });
+  //static final Map<String, MAJNode> map = {};
+  static Map<String, Map<String, MAJNode>> maps = {
+    defaultMapKey: {},
+  };
 
   /// adds a key and node reference to the map
   /// overwrites existing entries unless check == true
@@ -30,11 +26,12 @@ class MAJProvider with ChangeNotifier {
     required String path,
     required MAJNode node,
     bool check = false,
+    String mapKey = MAJProvider.defaultMapKey,
   }) {
     // check for overwrite throw error if will happen and check true
     if (check) {
-      for (int i = 0; i < map.keys.length; i++) {
-        if (map.keys.elementAt(i) == path) {
+      for (int i = 0; i < maps[mapKey]!.keys.length; i++) {
+        if (maps[mapKey]!.keys.elementAt(i) == path) {
           throw Exception(
             "MAJProvider: Cannot add '$path' to map, because it already exists",
           );
@@ -43,25 +40,38 @@ class MAJProvider with ChangeNotifier {
     }
 
     // add the node reference with path to the map
-    map[path] = node;
+    maps[mapKey]![path] = node;
   }
 
   /// removes a value from the map if the key exists
   /// otherwise null is returned
   static MAJNode? removeFromMap({
     required String path,
+    String mapKey = MAJProvider.defaultMapKey,
   }) {
-    return map.remove(path);
+    return maps[mapKey]!.remove(path);
   }
+
+  /// the default map key used if by MAJBuilder if no key is specified
+  final String mapKey;
+
+  /// the node to be displayed currently
+  MAJNode currentNode;
+
+  MAJProvider({
+    required this.currentNode,
+    this.mapKey = defaultMapKey,
+  });
 
   /// pass a valid path ex /root/node
   /// this will make the MAJBuilder change the currently displayed node to the
   /// node with the path passed
   /// ex:
   /// context.read<MAJProvider>().navigateTo("/root/node");
-  void navigateTo(String path) {
-    byPathElseByNode = true;
-    currentPath = path;
+  void navigateTo({
+    required String path,
+  }) {
+    currentNode = maps[mapKey]![path]!;
     notifyListeners();
   }
 
@@ -70,8 +80,9 @@ class MAJProvider with ChangeNotifier {
   /// to the passed node
   /// ex:
   /// context.read<MAJProvider>().navigateToByNode(myNode);
-  void navigateToByNode(MAJNode nodeTo) {
-    byPathElseByNode = false;
+  void navigateToByNode({
+    required MAJNode nodeTo,
+  }) {
     currentNode = nodeTo;
     notifyListeners();
   }